Fix #176 - Move ITs to specific modules and export test-jars
