+++ langchain4j/integration-tests/src/test/java/io/quarkiverse/flow/langchain4j/it/FlowAgentsBuilderIT.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.flow.langchain4j;
+package io.quarkiverse.flow.langchain4j.it;
 
 import static org.assertj.core.api.Assertions.assertThat;
