+++ langchain4j/pom.xml
@@ -20,6 +20,7 @@
     <modules>
         <module>deployment</module>
         <module>runtime</module>
+        <module>integration-tests</module>
     </modules>
 
 </project>
