+++ langchain4j/runtime/src/test/java/io/quarkiverse/flow/langchain4j/Agents.java
@@ -1,35 +1,12 @@
 package io.quarkiverse.flow.langchain4j;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import dev.langchain4j.agent.tool.Tool;
 import dev.langchain4j.agentic.Agent;
-import dev.langchain4j.agentic.declarative.ActivationCondition;
-import dev.langchain4j.agentic.declarative.ConditionalAgent;
-import dev.langchain4j.agentic.declarative.ExitCondition;
-import dev.langchain4j.agentic.declarative.LoopAgent;
-import dev.langchain4j.agentic.declarative.LoopCounter;
-import dev.langchain4j.agentic.declarative.Output;
-import dev.langchain4j.agentic.declarative.ParallelAgent;
 import dev.langchain4j.agentic.declarative.SequenceAgent;
-import dev.langchain4j.agentic.scope.AgenticScope;
-import dev.langchain4j.agentic.scope.ResultWithAgenticScope;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
 import io.quarkiverse.langchain4j.RegisterAiService;
 
 public class Agents {
-
-    static AtomicInteger loopCount;
-
-    public enum RequestCategory {
-        LEGAL,
-        MEDICAL,
-        TECHNICAL,
-        UNKNOWN
-    }
 
     @RegisterAiService
     public interface StoryCreatorWithConfigurableStyleEditor {
@@ -73,118 +50,6 @@ public class Agents {
         String editStory(@V("story") String story, @V("style") String style);
     }
 
-    @RegisterAiService
-    public interface EveningPlannerAgent {
-        @Output
-        static List<EveningPlan> createPlans(@V("movies") List<String> movies, @V("meals") List<String> meals) {
-            List<EveningPlan> moviesAndMeals = new ArrayList<>();
-            for (int i = 0; i < movies.size(); i++) {
-                if (i >= meals.size()) {
-                    break;
-                }
-                moviesAndMeals.add(new EveningPlan(movies.get(i), meals.get(i)));
-            }
-            return moviesAndMeals;
-        }
-
-        @ParallelAgent(outputKey = "plans", subAgents = { FoodExpert.class, MovieExpert.class })
-        List<EveningPlan> plan(@V("mood") String mood);
-    }
-
-    public interface FoodExpert {
-        @UserMessage("""
-                You are a great evening planner.
-                Propose a list of 3 meals matching the given mood.
-                The mood is {{mood}}.
-                For each meal, just give the name of the meal.
-                Provide a list with the 3 items and nothing else.
-                """)
-        @Agent(outputKey = "meals")
-        List<String> findMeal(@V("mood") String mood);
-    }
-
-    public interface MovieExpert {
-        @UserMessage("""
-                You are a great evening planner.
-                Propose a list of 3 movies matching the given mood.
-                The mood is {{mood}}.
-                Provide a list with the 3 items and nothing else.
-                """)
-        @Agent(outputKey = "movies")
-        List<String> findMovie(@V("mood") String mood);
-    }
-
-    public interface CategoryRouter {
-        @UserMessage("""
-                Analyze the following user request and categorize it as 'legal', 'medical' or 'technical'.
-                In case the request doesn't belong to any of those categories categorize it as 'unknown'.
-                Reply with only one of those words and nothing else.
-                The user request is: '{{request}}'.
-                """)
-        @Agent(description = "Categorize a user request", outputKey = "category")
-        RequestCategory classify(@V("request") String request);
-    }
-
-    public interface MedicalExpert {
-        @UserMessage("""
-                You are a medical expert.
-                Analyze the following user request under a medical point of view and provide the best possible answer.
-                The user request is {{request}}.
-                """)
-        @Tool("A medical expert")
-        @Agent(description = "A medical expert", outputKey = "response")
-        String medical(@V("request") String request);
-    }
-
-    public interface LegalExpert {
-        @UserMessage("""
-                You are a legal expert.
-                Analyze the following user request under a legal point of view and provide the best possible answer.
-                The user request is {{request}}.
-                """)
-        @Tool("A legal expert")
-        @Agent(description = "A legal expert", outputKey = "response")
-        String legal(@V("request") String request);
-    }
-
-    public interface TechnicalExpert {
-        @UserMessage("""
-                You are a technical expert.
-                Analyze the following user request under a technical point of view and provide the best possible answer.
-                The user request is {{request}}.
-                """)
-        @Tool("A technical expert")
-        @Agent(description = "A technical expert", outputKey = "response")
-        String technical(@V("request") String request);
-    }
-
-    @RegisterAiService
-    public interface ExpertsAgent {
-
-        @ActivationCondition(MedicalExpert.class)
-        static boolean activateMedical(@V("category") RequestCategory category) {
-            return category == RequestCategory.MEDICAL;
-        }
-
-        @ActivationCondition(LegalExpert.class)
-        static boolean activateLegal(AgenticScope agenticScope) {
-            return agenticScope.readState("category", RequestCategory.UNKNOWN) == RequestCategory.LEGAL;
-        }
-
-        @ActivationCondition(TechnicalExpert.class)
-        static boolean activateTechnical(@V("category") RequestCategory category) {
-            return category == RequestCategory.TECHNICAL;
-        }
-
-        @ConditionalAgent(outputKey = "response", subAgents = { MedicalExpert.class, LegalExpert.class, TechnicalExpert.class })
-        String askExpert(@V("request") String request);
-    }
-
-    public interface ExpertRouterAgent {
-        @SequenceAgent(outputKey = "response", subAgents = { CategoryRouter.class, ExpertsAgent.class })
-        ResultWithAgenticScope<String> ask(@V("request") String request);
-    }
-
     public interface DumbAgent {
         @Agent(description = "A dumb agent to workaround a dumb check", outputKey = "loopCounter")
         int dumb(@V("request") String request);
@@ -205,41 +70,4 @@ public class Agents {
         String dumb4();
 
     }
-
-    public interface StyleScorer {
-        @UserMessage("""
-                You are a critical reviewer.
-                Give a review score between 0.0 and 1.0 for the following story based on how well it aligns with the style '{{style}}'.
-                Return only the score and nothing else.
-
-                The story is: "{{story}}"
-                """)
-        @Agent(description = "Score a story based on how well it aligns with a given style", outputKey = "score")
-        double scoreStyle(@V("story") String story, @V("style") String style);
-    }
-
-    @RegisterAiService
-    public interface StyleReviewLoopAgentWithCounter {
-
-        @ExitCondition(testExitAtLoopEnd = true)
-        static boolean exit(@V("score") double score, @LoopCounter int loopCounter) {
-            loopCount.set(loopCounter);
-            return score >= 0.8;
-        }
-
-        @LoopAgent(description = "Review the given story to ensure it aligns with the specified style", outputKey = "story", maxIterations = 5, subAgents = {
-                StyleScorer.class, StyleEditor.class })
-        String write(@V("story") String story);
-    }
-
-    @RegisterAiService
-    public interface StoryCreatorWithReviewWithCounter {
-        @SequenceAgent(outputKey = "story", subAgents = { CreativeWriter.class, StyleReviewLoopAgentWithCounter.class })
-        @Agent("write")
-        ResultWithAgenticScope<String> write(@V("topic") String topic, @V("style") String style);
-    }
-
-    public record EveningPlan(String movie, String meal) {
-    }
-
 }
